+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.1.0
 - Add support for JSON comments in _variables_ and in JSON variable files and environment variables
 - Add support for multiple glob patterns separeted by a semi-colon (`;`) using [fast-glob](https://github.com/mrmlnc/fast-glob) in variable file paths
+- Add support for YAML variable files (`.yml` or `.yaml`)
 
 ## v1.0.0
 - Initial Release of the ReplaceTokens action