+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add support for JSON comments in _variables_ and in JSON variable files and environment variables
 - Add support for multiple glob patterns separeted by a semi-colon (`;`) using [fast-glob](https://github.com/mrmlnc/fast-glob) in variable file paths
 - Add support for YAML variable files (`.yml` or `.yaml`)
+- Fix log level (`info`) for groups
 
 ## v1.0.0
 - Initial Release of the ReplaceTokens action