+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add support for multiple glob patterns separeted by a semi-colon (`;`) using [fast-glob](https://github.com/mrmlnc/fast-glob) in variable file paths
 - Add support for YAML variable files (`.yml` or `.yaml`)
 - Fix log level (`info`) for groups
+- Add anonymous telemetry usage
 
 ## v1.0.0
 - Initial Release of the ReplaceTokens action