+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v1.1.1
+- Fix variable case-sensitivity ([#7](https://github.com/qetza/replacetokens-action/issues/7)).
+
 ## v1.1.0
 - Add support for JSON comments in _variables_ and in JSON variable files and environment variables
 - Add support for multiple glob patterns separeted by a semi-colon (`;`) using [fast-glob](https://github.com/mrmlnc/fast-glob) in variable file paths